moved OAuthUserInfo to service jar, so it can be reused
introduced class for dormitory
