+++ sch-pek-ejb-services/src/main/java/hu/sch/services/dto/OAuthDormitory.java
@@ -1,0 +1,28 @@
+package hu.sch.services.dto;
+
+import com.google.gson.annotations.SerializedName;
+
+public class OAuthDormitory {
+
+    @SerializedName("buildingName")
+    private String buildingName;
+
+    @SerializedName("roomNumber")
+    private String room;
+
+    public String getBuildingName() {
+        return buildingName;
+    }
+
+    public void setBuildingName(String buildingName) {
+        this.buildingName = buildingName;
+    }
+
+    public String getRoom() {
+        return room;
+    }
+
+    public void setRoom(String room) {
+        this.room = room;
+    }
+}
